Move Geocode out of ProcessOSM
